# api/docs/swagger.yaml
info:
  contact: {}
paths:
...
  /api/v1/admin/add:
    post:
      consumes:
      - application/json
      description: Add an API key.
      parameters:
      - description: Authentication token (e.g., Bearer <token>)
        in: header
        name: Authorization
        required: true
        type: string
      - description: key
        in: query
        name: key
        type: string
      - description: scope
        in: query
        name: scope
        type: string
      - description: permissions
        in: query
        name: permissions
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - admin
  /api/v1/admin/delete:
    post:
      consumes:
      - application/json
      description: Delete an API key.
      parameters:
      - description: Authentication token (e.g., Bearer <token>)
        in: header
        name: Authorization
        required: true
        type: string
      - description: key
        in: query
        name: key
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - admin
  /api/v1/admin/edit:
    post:
      consumes:
      - application/json
      description: Edit an API key.
      parameters:
      - description: Authentication token (e.g., Bearer <token>)
        in: header
        name: Authorization
        required: true
        type: string
      - description: key
        in: query
        name: key
        type: string
      - description: scope
        in: query
        name: scope
        type: string
      - description: permissions
        in: query
        name: permissions
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - admin
...
swagger: "2.0"